search: test indexedSearchRequest
We adapt the existing test for TestZoektSearchHEAD to instead test the
new indexedSearchRequest abstraction. This will allow us to test the
interaction between zoektIndexedRepos and zoektIndexedSearch.

The test code has had a few changes to it, but is the same
functionality. Additionally we test the returned file URLs (an important
component of what zoektSearch does).

We also remove the test for zoekt returning an error. Allowing this test
makes the test cases harder to read, and doesn't provide that much
value.